backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.9.0
python-multipart==0.0.6
scikit-learn>=1.5.0
torch>=2.1.0
transformers>=4.35.0
sentence-transformers>=2.2.2
pandas>=2.1.3
numpy>=1.26.2
networkx>=3.2.1
shap>=0.43.0
lime>=0.2.0.1
PyPDF2>=3.0.1
python-docx>=1.1.0
nltk>=3.8.1
imbalanced-learn>=0.11.0
joblib>=1.3.2
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=7.4.3
httpx>=0.25.2
